Dev/nasiima

+++ app/models.py
@@ -9,7 +9,7 @@ class Users(db.Model):
     name = db.Column(db.String(64))
     username = db.Column(db.String(100))
     email = db.Column(db.String(100))
-    dob = db.Column(db.Integer)
+    dob = db.Column(db.Float)
     password_digest = db.Column(db.String(10000))
     preferences =  db.Column(db.String(140))
     picture = db.Column(db.String(10000))
